+++ src/Transition/Core/Language.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Copyright © MB Arbatos Klubas. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace FreshAdvance\Invoice\Transition\Core;
+
+/**
+ * @mixin \OxidEsales\Eshop\Core\Language
+ */
+class Language extends Language_parent
+{
+    protected $_iTplLanguageId = null;
+
+    public function faForceSetTplLanguage(int $languageId): void
+    {
+        $this->_iTplLanguageId = $languageId;
+    }
+}