+++ tests/PhpStan/phpstan-bootstrap.php
@@ -11,3 +11,8 @@ class_alias(
     \OxidEsales\Eshop\Application\Model\OrderArticle::class,
     \FreshAdvance\Invoice\Transition\Model\OrderArticle_parent::class
 );
+
+class_alias(
+    \OxidEsales\Eshop\Core\Language::class,
+    \FreshAdvance\Invoice\Transition\Core\Language_parent::class
+);