+++ src/Document/MpdfDocument/MpdfFactory.php
@@ -9,16 +9,20 @@ declare(strict_types=1);
 
 namespace FreshAdvance\Invoice\Document\MpdfDocument;
 
+use FreshAdvance\Invoice\Service\ModuleSettingsInterface;
 use Mpdf\Mpdf;
 
 class MpdfFactory
 {
     public function __construct(
+        private ModuleSettingsInterface $moduleSettings
     ) {
     }
 
     public function create(): Mpdf
     {
-        return new Mpdf();
+        return new Mpdf([
+            'PDFA' => $this->moduleSettings->isForArchive()
+        ]);
     }
 }