+++ src/Service/ModuleSettings.php
@@ -16,6 +16,7 @@ class ModuleSettings implements ModuleSettingsInterface
 {
     public const SETTING_DOCUMENT_FOOTER = 'fa_invoice_DocumentFooter';
     public const SETTING_DOCUMENT_FILENAME_PREFIX = 'fa_invoice_FilenamePrefix';
+    public const SETTING_DOCUMENT_IS_FOR_ARCHIVE = 'fa_invoice_IsForArchive';
 
     public function __construct(
         private ModuleSettingServiceInterface $moduleSettingService
@@ -30,6 +31,14 @@ class ModuleSettings implements ModuleSettingsInterface
     public function getFilePrefix(): string
     {
         return $this->getStringSetting(self::SETTING_DOCUMENT_FILENAME_PREFIX);
+    }
+
+    public function isForArchive(): bool
+    {
+        return $this->moduleSettingService->getBoolean(
+            self::SETTING_DOCUMENT_IS_FOR_ARCHIVE,
+            Module::MODULE_ID
+        );
     }
 
     private function getStringSetting(string $key): string