+++ src/Service/ModuleSettingsInterface.php
@@ -7,4 +7,6 @@ interface ModuleSettingsInterface
     public function getDocumentFooter(): string;
 
     public function getFilePrefix(): string;
+
+    public function isForArchive(): bool;
 }