+++ tests/Integration/Document/MpdfDocument/MpdfFactoryTest.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace Document\MpdfDocument;
 
 use FreshAdvance\Invoice\Document\MpdfDocument\MpdfFactory;
+use FreshAdvance\Invoice\Service\ModuleSettingsInterface;
 use Mpdf\Mpdf;
 use PHPUnit\Framework\TestCase;
 
@@ -20,7 +21,22 @@ class MpdfFactoryTest extends TestCase
 {
     public function testCreate(): void
     {
-        $sut = new MpdfFactory();
+        $sut = new MpdfFactory(
+            moduleSettings: $this->createStub(ModuleSettingsInterface::class)
+        );
         $this->assertInstanceOf(Mpdf::class, $sut->create());
+    }
+
+    public function testArchiveFlagConfigured(): void
+    {
+        $settingsStub = $this->createMock(ModuleSettingsInterface::class);
+        $settingsStub->method('isForArchive')->willReturn(true);
+
+        $sut = new MpdfFactory(
+            moduleSettings: $settingsStub
+        );
+        $object = $sut->create();
+
+        $this->assertSame(true, $object->PDFA);
     }
 }