+++ metadata.php
@@ -21,7 +21,7 @@ $aModule = [
     'description'  => [
         'en' => 'Invoice module for OXID eShop.',
     ],
-    'version'     => '1.3.1',
+    'version'     => '2.0.0',
     'author'       => 'Anton Fedurtsya',
     'email'        => 'anton@fedurtsya.com',
     'url'         => '',